serde: tag_invoke for std::chrono types
