+++ src/v/serde/serde.h
@@ -162,59 +162,6 @@ inline constexpr auto const is_serde_compatible_v
     || is_std_unordered_map<T>
     || is_fragmented_vector<T> || is_chunked_fifo<T> || reflection::is_tristate<T> || std::is_same_v<T, ss::net::inet_address>;
 
-template<class R, class P>
-int64_t checked_duration_cast_to_nanoseconds(
-  const std::chrono::duration<R, P>& duration) {
-    static_assert(
-      __has_builtin(__builtin_mul_overflow),
-      "__builtin_mul_overflow not supported.");
-    using nano_period = std::chrono::nanoseconds::period;
-    using nano_rep = typename std::chrono::nanoseconds::rep;
-    // This is how duration_cast determines the output type.
-    using output_type = typename std::common_type<R, nano_rep, intmax_t>::type;
-    using ratio = std::ratio_divide<P, nano_period>;
-
-    static_assert(
-      std::is_same_v<output_type, nano_rep>,
-      "Output type is not the same rep as std::chrono::nanoseconds");
-
-    // Extra check to ensure the output type is same as the underlying type
-    // supported in lib[std]c++.
-    static_assert(
-      std::is_same_v<output_type, int64_t>
-        || std::is_same_v<output_type, long long>,
-      "Output type not in supported integer types.");
-
-    constexpr auto ratio_num = static_cast<output_type>(ratio::num);
-    const auto dur = static_cast<output_type>(duration.count());
-    output_type mul;
-    if (unlikely(__builtin_mul_overflow(dur, ratio_num, &mul))) {
-        // Clamp the value.
-        // Log here to ensure that it is picked up by duck tape. Ideally this
-        // should never happen, but in case it does, we have a log trail.
-        //
-        // If you are here because your ducktape test failed with this
-        // BadLogLine check, it means that you serialized a duration type that
-        // caused an overflow when casting to nanoseconds. Clamp your duration
-        // to nanoseconds::max().
-        using input_type = typename std::chrono::duration<R, P>;
-        vlog(
-          serde_log.error,
-          "Overflow or underflow detected when casting to nanoseconds, "
-          "clamping to a limit. Input: {}  type: {}",
-          duration.count(),
-          type_str<input_type>());
-        return duration.count() < 0 ? std::chrono::nanoseconds::min().count()
-                                    : std::chrono::nanoseconds::max().count();
-    }
-    // No overflow/underflow detected, we are safe to cast.
-    return std::chrono::duration_cast<std::chrono::nanoseconds>(duration)
-      .count();
-}
-
-template<typename Rep, typename Period>
-void write(iobuf& out, std::chrono::duration<Rep, Period> t);
-
 inline void write(iobuf& out, ss::net::inet_address t);
 
 template<typename T, typename Tag, typename IsConstexpr>
@@ -253,31 +200,6 @@ requires is_envelope<std::decay_t<T>>
 void write(iobuf& out, T t);
 
 inline void write(iobuf& out, iobuf t);
-
-template<typename Rep, typename Period>
-void write(iobuf& out, std::chrono::duration<Rep, Period> t) {
-    // We explicitly serialize it as ns to avoid any surprises like
-    // seastar updating underlying duration types without
-    // notice. See https://github.com/redpanda-data/redpanda/pull/5002
-    //
-    // Check for overflows/underflows.
-    // For ex: a millisecond and nanosecond use the same underlying
-    // type int64_t but converting from one to other can easily overflow,
-    // this is by design.
-    // Since we serialize with ns precision, there is a restriction of
-    // nanoseconds::max()'s equivalent on the duration to be serialized.
-    // On a typical platform which uses int64_t for 'rep', it roughly
-    // translates to ~292 years.
-    //
-    // If we detect an overflow, we will clamp it to maximum supported
-    // duration, which is nanosecond::max() and if there is an underflow,
-    // we clamp it to minimum supported duration which is nanosecond::min().
-    static_assert(
-      !std::is_floating_point_v<Rep>,
-      "Floating point duration conversions are prone to precision and "
-      "rounding issues.");
-    write<int64_t>(out, checked_duration_cast_to_nanoseconds(t));
-}
 
 inline void write(iobuf& out, ss::net::inet_address t) {
     iobuf address_bytes;
@@ -460,14 +382,6 @@ inline void write(iobuf& out, iobuf t) {
     out.append(t.share(0, t.size_bytes()));
 }
 
-template<typename Clock, typename Duration>
-void write(iobuf&, std::chrono::time_point<Clock, Duration> t) {
-    static_assert(
-      !is_chrono_time_point<decltype(t)>,
-      "Time point serialization is risky and can have unintended "
-      "consequences. Check with Redpanda team before fixing this.");
-}
-
 template<typename T>
 std::decay_t<T> read_nested(iobuf_parser&, std::size_t bytes_left_limit);
 
@@ -487,10 +401,6 @@ std::decay_t<T> read(iobuf_parser& in) {
 template<typename T>
 void read_nested(iobuf_parser& in, T& t, std::size_t const bytes_left_limit) {
     using Type = std::decay_t<T>;
-    static_assert(
-      !is_chrono_time_point<Type>,
-      "Time point serialization is risky and can have unintended "
-      "consequences. Check with Redpanda team before fixing this.");
     static_assert(has_serde_read<T> || is_serde_compatible_v<Type>);
 
     if constexpr (is_envelope<Type>) {
@@ -618,13 +528,6 @@ void read_nested(iobuf_parser& in, T& t, std::size_t const bytes_left_limit) {
             t.push_back(read_nested<value_type>(in, bytes_left_limit));
         }
         t.shrink_to_fit();
-    } else if constexpr (is_chrono_duration<Type>) {
-        static_assert(
-          !std::is_floating_point_v<typename Type::rep>,
-          "Floating point duration conversions are prone to precision and "
-          "rounding issues.");
-        auto rep = read_nested<int64_t>(in, bytes_left_limit);
-        t = std::chrono::duration_cast<Type>(std::chrono::nanoseconds{rep});
     } else if constexpr (reflection::is_tristate<T>) {
         int8_t flag = read_nested<int8_t>(in, bytes_left_limit);
         if (flag == -1) {
@@ -876,6 +779,7 @@ inline serde::serde_size_t peek_body_size(iobuf_parser& in) {
 
 } // namespace serde
 #include "serde/rw/bool_class.h"
+#include "serde/rw/chrono.h"
 #include "serde/rw/enum.h"
 #include "serde/rw/scalar.h"
 #include "serde/rw/sstring.h"
